+++ slices.go
@@ -32,3 +32,13 @@ func difference(a, b []string) []string {
 	}
 	return diff
 }
+
+// contains tells whether a contains x.
+func contains(a []string, x string) bool {
+	for _, n := range a {
+		if x == n {
+			return true
+		}
+	}
+	return false
+}